add wait for job option to update config
